stdlib: move 'String : Hashable' conformance to a separate file
I'm about to add more code around it.
